Moved display.
